Make GH repos linkable
