+++ querying/github.go
@@ -2,6 +2,7 @@ package querying
 
 import (
 	"context"
+	"strings"
 	"sync"
 
 	"golang.org/x/oauth2"
@@ -10,6 +11,8 @@ import (
 	"github.com/underdog-tech/vulnbot/configs"
 	"github.com/underdog-tech/vulnbot/logger"
 )
+
+const internalTopicKeyword = "internal"
 
 type githubClient interface {
 	Query(context.Context, interface{}, map[string]interface{}) error
@@ -36,55 +39,6 @@ func NewGithubDataSource(conf *configs.Config) GithubDataSource {
 		conf:     conf,
 		ctx:      context.Background(),
 	}
-}
-
-type githubVulnerability struct {
-	SecurityAdvisory struct {
-		Description string
-		Identifiers []struct {
-			Type  string
-			Value string
-		}
-	}
-	SecurityVulnerability struct {
-		Severity string
-		Package  struct {
-			Ecosystem string
-			Name      string
-		}
-	}
-}
-
-type orgRepo struct {
-	Name                string
-	Url                 string
-	VulnerabilityAlerts struct {
-		TotalCount int
-		PageInfo   struct {
-			EndCursor   githubv4.String
-			HasNextPage bool
-		}
-		Nodes []githubVulnerability
-	} `graphql:"vulnerabilityAlerts(states: OPEN, first: 100, after: $alertCursor)"`
-}
-
-type repositoryQuery struct {
-	Repository orgRepo `graphql:"repository(name: $repoName, owner: $orgName)"`
-}
-
-type orgVulnerabilityQuery struct {
-	Organization struct {
-		Name         string
-		Login        string
-		Repositories struct {
-			TotalCount int
-			PageInfo   struct {
-				EndCursor   githubv4.String
-				HasNextPage bool
-			}
-			Nodes []orgRepo
-		} `graphql:"repositories(orderBy: {field: NAME, direction: ASC}, isFork: false, isArchived: false, first: 100, after: $repoCursor)"`
-	} `graphql:"organization(login: $login)"`
 }
 
 // Ref: https://docs.github.com/en/graphql/reference/enums#securityadvisoryecosystem
@@ -150,7 +104,7 @@ func (gh *GithubDataSource) processRepoFindings(projects *ProjectCollection, rep
 
 	// Link directly to Dependabot findings.
 	// There doesn't appear to be a GraphQL property for this link.
-	project.Links["GitHub"] = repo.Url + "/security/dependabot"
+	project.Link = repo.Url + "/security/dependabot"
 
 	log.Debug().Str("project", project.Name).Msg("Processing findings for project.")
 
@@ -198,38 +152,6 @@ func (gh *GithubDataSource) processRepoFindings(projects *ProjectCollection, rep
 	return nil
 }
 
-type orgTeam struct {
-	Name         string
-	Slug         string
-	Repositories struct {
-		PageInfo struct {
-			EndCursor   githubv4.String
-			HasNextPage bool
-		}
-		Edges []struct {
-			Permission string
-			Node       struct {
-				Name       string
-				IsFork     bool
-				IsArchived bool
-			}
-		}
-	} `graphql:"repositories(orderBy: {field: NAME, direction: ASC}, first: 100, after: $repoCursor)"`
-}
-
-type orgRepoOwnerQuery struct {
-	Organization struct {
-		Teams struct {
-			TotalCount int
-			PageInfo   struct {
-				EndCursor   githubv4.String
-				HasNextPage bool
-			}
-			Nodes []orgTeam
-		} `graphql:"teams(orderBy: {field: NAME, direction: ASC}, first: 100, after: $teamCursor)"`
-	} `graphql:"organization(login: $login)"`
-}
-
 func (gh *GithubDataSource) gatherRepoOwners(projects *ProjectCollection) {
 	var ownerQuery orgRepoOwnerQuery
 	log := logger.Get()
@@ -253,7 +175,7 @@ func (gh *GithubDataSource) gatherRepoOwners(projects *ProjectCollection) {
 			}
 			// TODO: Handle pagination of repositories owned by a team
 			for _, repo := range team.Repositories.Edges {
-				if repo.Node.IsArchived || repo.Node.IsFork {
+				if repo.Node.IsArchived || repo.Node.IsFork || hasInternalTopic(repo.Node.RepositoryTopics) {
 					log.Debug().Str("Repo", repo.Node.Name).Bool("IsFork", repo.Node.IsFork).Bool("IsArchived", repo.Node.IsArchived).Msg("Skipping untracked repository.")
 					continue
 				}
@@ -271,4 +193,14 @@ func (gh *GithubDataSource) gatherRepoOwners(projects *ProjectCollection) {
 		}
 		queryVars["teamCursor"] = githubv4.NewString(ownerQuery.Organization.Teams.PageInfo.EndCursor)
 	}
+}
+
+// Function to check if the repository has "internal" in its topics
+func hasInternalTopic(repoTopics repositoryTopics) bool {
+	for _, edge := range repoTopics.Edges {
+		if strings.Contains(strings.ToLower(edge.Node.Topic.Name), internalTopicKeyword) {
+			return true
+		}
+	}
+	return false
 }
